# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Web Umkm",
        "description": "This For API System,And You Can Try It",
        "contact": {
            "email": "muhammadalghifari4321@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/barangtoko": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Looking Data Barang And Toko",
                "description": "Looking Data Barang And Toko",
                "operationId": "getBarangToko",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Insert Data BarangToko Admin",
                "description": "Insert Data BarangToko Admin",
                "operationId": "insertBarangToko",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "barang_id": {},
                                    "toko_id": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/barangtoko/{id}": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Get By Id Data Barang And Toko",
                "description": "Get By Id Data Barang And Toko",
                "operationId": "getByIdBarangToko",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Get Data And Relation",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Update Data BarangToko Admin",
                "description": "Update Data BarangToko Admin",
                "operationId": "updateBarangToko",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Update",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "barang_id": {},
                                    "toko_id": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Delete Data BarangToko Admin",
                "description": "Delete Data BarangToko Admin",
                "operationId": "deleteBarangToko",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Delete",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    ]
}